# [GTK] Speedup TreeItem retrieval by removing redundant iterations #882