--- a/tests/org.eclipse.swt.tests/JUnit Tests/org/eclipse/swt/tests/junit/Test_org_eclipse_swt_widgets_TreeItem.java
+++ b/tests/org.eclipse.swt.tests/JUnit Tests/org/eclipse/swt/tests/junit/Test_org_eclipse_swt_widgets_TreeItem.java
@@ -1174,6 +1174,15 @@ public void test_setTextILjava_lang_String(){
 
 }
 
+@Test
+public void test_removeAll() {
+	TreeItem item = new TreeItem(treeItem, SWT.NONE);
+	assertEquals(1, treeItem.getItemCount());
+	treeItem.removeAll();
+	assertEquals(0, treeItem.getItemCount());
+	assertTrue(item.isDisposed());
+}
+
 /* custom */
 TreeItem treeItem;
 Tree tree;