@SyntevoAlex 2023-12-03
I can review it a few days later.
Feel free to find another reviewer if you don't want to wait. But yes, it's not going to be too easy