@SyntevoAlex 2023-12-05
I feel that https://github.com/eclipse-platform/eclipse.platform.swt/issues/930 is significantly more important, so I'll have to handle it first.