@SyntevoAlex 2024-01-07
This change makes sense to me.
It's somewhat non-trivial, especially around removing `SWT.VIRTUAL` branch, but it looks correct to me.